backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.20
azure-storage-blob==12.24.0
python-dotenv==1.0.1
httpx==0.28.1
gunicorn==23.0.0
